Refactored again moving controllers out of routes sub folder and into controllers class
